# REST APIs Implementation/api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Film Manager",
    "description": "REST APIs for the interaction with the Film Manager service.",
    "termsOfService": "http://swagger.io/terms/",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "http://localhost:3001/"
    }
  ],
...
  "paths": {
...
    "/api/films/public/{filmId}/reviews/{reviewerId}/editRequests": {
      "post": {
        "tags": [
          "api/films/public/{filmId}/reviews/{reviewerId}/editRequests"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Request an edit to a review",
        "description": "The reviewer with ID reviewerId requests an edit to the review of the film with ID filmId. This operation can only be performed by the reviewer.",
        "operationId": "requestEditReview",
        "parameters": [
          {
            "name": "filmId",
            "in": "path",
            "description": "ID of the film whose review must be edited",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "reviewerId",
            "in": "path",
            "description": "ID of the user to whom the review has been issued",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "The deadline of the edit request",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "deadline": {
                    "type": "string",
                    "format": "date-time-timezone",
                    "pattern": "^([0-9]{4})-(0[1-9]|1[0-2])-(0[1-9]|[12][0-9]|3[01])T([01][0-9]|2[0-3]):([0-5][0-9])(Z|[+-]([01][0-9]|2[0-3]):[0-5][0-9])$",
                    "example": "2025-10-01T12:00Z"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditRequest"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {}
          },
          "401": {
            "description": "Unauthorized",
            "content": {}
          },
          "403": {
            "description": "Forbidden (the user is not the reviewer)",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {}
          },
          "409": {
            "description": "Conflict",
            "content": {}
          },
          "500": {
            "description": "Internal Server Error",
            "content": {}
          }
        }
      },
      "get": {
        "tags": [
          "api/films/public/{filmId}/reviews/{reviewerId}/editRequests"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Retrieve the edit requests for a review",
        "description": "The edit requests for the review of the film with ID filmId and issued to the user with ID reviewerId are retrieved.",
        "operationId": "getEditRequests",
        "parameters": [
          {
            "name": "filmId",
            "in": "path",
            "description": "ID of the film whose review must be edited",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "reviewerId",
            "in": "path",
            "description": "ID of the user to whom the review has been issued",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditRequests"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {}
          },
          "403": {
            "description": "Forbidden",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {}
          },
          "500": {
            "description": "Internal Server Error",
            "content": {}
          }
        }
      }
    },
    "/api/films/public/{filmId}/reviews/{reviewerId}/editRequests/{requestId}": {
      "get": {
        "tags": [
          "api/films/public/{filmId}/reviews/{reviewerId}/editRequests/{requestId}"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Retrieve an edit request",
        "description": "The edit request with ID requestId for the review of the film with ID filmId and issued to the user with ID reviewerId is retrieved. This operation can only be performed by the reviewer.",
        "operationId": "getSingleEditRequest",
        "parameters": [
          {
            "name": "filmId",
            "in": "path",
            "description": "ID of the film whose review must be edited",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "reviewerId",
            "in": "path",
            "description": "ID of the user to whom the review has been issued",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "requestId",
            "in": "path",
            "description": "ID of the edit request",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditRequest"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {}
          },
          "401": {
            "description": "Unauthorized",
            "content": {}
          },
          "403": {
            "description": "Forbidden (user is not the reviewer or the owner)",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {}
          },
          "500": {
            "description": "Internal Server Error",
            "content": {}
          }
        }
      },
      "put": {
        "tags": [
          "api/films/public/{filmId}/reviews/{reviewerId}/editRequests/{requestId}"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Update an edit request",
        "description": "The edit request with ID requestId for the review of the film with ID filmId and issued to the user with ID reviewerId is updated. This operation can only be performed by the film owner.",
        "operationId": "updateEditRequest",
        "parameters": [
          {
            "name": "filmId",
            "in": "path",
            "description": "ID of the film whose review must be edited",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "reviewerId",
            "in": "path",
            "description": "ID of the user to whom the review has been issued",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "requestId",
            "in": "path",
            "description": "ID of the edit request",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "The response to the edit request",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "accepted": {
                    "type": "boolean",
                    "enum": [
                      true,
                      false
                    ],
                    "example": true
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content",
            "content": {}
          },
          "400": {
            "description": "Bad Request",
            "content": {}
          },
          "401": {
            "description": "Unauthorized",
            "content": {}
          },
          "403": {
            "description": "Forbidden (the user is not the owner)",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {}
          },
          "409": {
            "description": "Conflict",
            "content": {}
          },
          "500": {
            "description": "Internal Server Error",
            "content": {}
          }
        }
      },
      "delete": {
        "tags": [
          "api/films/public/{filmId}/reviews/{reviewerId}/editRequests/{requestId}"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Delete an edit request",
        "description": "The edit request with ID requestId for the review of the film with ID filmId and issued to the user with ID reviewerId is deleted. This operation can only be performed by the reviewer.",
        "operationId": "deleteEditRequest",
        "parameters": [
          {
            "name": "filmId",
            "in": "path",
            "description": "ID of the film whose review must be edited",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "reviewerId",
            "in": "path",
            "description": "ID of the user to whom the review has been issued",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "requestId",
            "in": "path",
            "description": "ID of the edit request",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful operation - No Content",
            "content": {}
          },
          "400": {
            "description": "Bad Request",
            "content": {}
          },
          "401": {
            "description": "Unauthorized",
            "content": {}
          },
          "403": {
            "description": "Forbidden (the user is not the reviewer)",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {}
          },
          "500": {
            "description": "Internal Server Error",
            "content": {}
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EditRequest": {
        "description": "schema for edit request data structures",
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "type": "string",
            "description": "path to the JSON Schema to use for the validation of the JSON file describing the edit request"
          },
          "id": {
            "type": "integer",
            "description": "unique identifier of the edit request"
          },
          "filmId": {
            "type": "integer",
            "description": "unique identifier of the film for which a edit review request has been issued"
          },
          "reviewerId": {
            "type": "integer",
            "description": "unique identifier of the the user who has send the edit review request"
          },
          "deadline": {
            "type": "string",
            "format": "date-time-timezone",
            "pattern": "^([0-9]{4})-(0[1-9]|1[0-2])-(0[1-9]|[12][0-9]|3[01])T([01][0-9]|2[0-3]):([0-5][0-9])(Z|[+-]([01][0-9]|2[0-3]):[0-5][0-9])$",
            "description": "the date-time-timezone when the request for edit the review will be expired"
          },
          "status": {
            "type": "string",
            "description": "the status of the request",
            "enum": [
              "pending",
              "accepted",
              "rejected"
            ],
            "default": "pending"
          },
          "self": {
            "type": "string",
            "format": "uri"
          }
        },
        "required": [
          "id",
          "filmId",
          "reviewerId",
          "deadline"
        ]
      },
      "EditRequests": {
        "description": "schema for editRequests data structures",
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "type": "string",
            "description": "path to the JSON Schema to use for the validation of the JSON data structure describing the editRequests"
          },
          "totalPages": {
            "type": "integer",
            "minimum": 0,
            "description": "total number of pages"
          },
          "currentPage": {
            "type": "integer",
            "minimum": 0,
            "description": "the page included in this object"
          },
          "totalItems": {
            "type": "integer",
            "minimum": 0,
            "description": "the total number of items in the current page"
          },
          "editRequests": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EditRequest"
            },
            "description": "the edit requests of the current page"
          },
          "next": {
            "type": "string",
            "format": "uri"
          }
        },
        "required": [
          "totalPages"
        ]
      }
    },
    "securitySchemes": {
      "cookieAuth": {
        "description": "JWT to identify the authenticated user.",
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid"
      }
    }
  }
}